+++ Proyecto Intermodular/api/models/TicketModel.cs
@@ -7,7 +7,7 @@ namespace Proyecto_Intermodular.api.models
     public class TicketModel
     {
         private string id;
-        private float total;
+        private double total;
         private string text;
         private bool isPaid;
         private string createdAt;
@@ -26,7 +26,7 @@ namespace Proyecto_Intermodular.api.models
         }
 
         public string Id { get => id; set => id = value; }
-        public float Total { get => total; set => total = value; }
+        public double Total { get => total; set => total = value; }
         public string Text { get => text; set => text = value; }
         public bool IsPaid { get => isPaid; set => isPaid = value; }
         public string CreatedAt { get => createdAt; set => createdAt = value; }
